refactor(openfold3): own structure prediction contract
